Driver code
The main code which instructs the conversion of natural language query to a query suitable for input on the database containing network traffic and statistics. 

Takes natural language query. Converts to a suitable query. Query Fired on network database. Result in a data frame converted to a natural language format and sent back to the network admin (end user).